+++ README.md
@@ -1,2 +1,8 @@
 # Antares's Fenrir Telegram Bot
 yeeeet
+
+### Running
+
+```bash
+$ python -m src.main
+```
