+++ src/db/__init.py
@@ -1,0 +1,3 @@
+"""
+Database modules
+"""
