+++ src/handlers/__init__.py
@@ -1,0 +1,3 @@
+"""
+Telegram handlers
+"""
