+++ src/models/__init__.py
@@ -1,0 +1,3 @@
+"""
+Data models
+"""
